add: adding and organizing functions
The added functions are
- InputTools.check_param -> Confirmation of input file parameters
- InputTools.Change.model_parameter -> Change *model_parameter*
- InputTools.Change.simulation_parameter -> Change *simulation_parameter*
- InputTools.Change.material_property -> Change *material_property*
- InputTools.Change.output_parameter -> Change *output_parameter*
These functions are intended for external use.

The added internal-functions are
- _Tools.print_param -> For checking parameter
- _Tools.change_routine -> For changing parameter
- _Tools.change_value -> For changing parameter
- _Tools.read_data -> For changing parameter
- _Tools.rewrite_data -> For changing parameter
These are internal functions that are assumed to be used in the aforementioned functions.

Also, some functions have been moved for the sake of organization, and there are
- InputTools.MkInputFile.outof_package
- InputTools.MkInputFile.into_package
- _Tools.copy_fmtfile
- _Tools.mk_fmtfile
- _Tools.mk_setting_dict

These modifications and additions are features related to the input file.
These are for convenience and basically have no effect on the simulation itself.